add passive mode

+++ handlers/msgHandler.js
@@ -6,7 +6,8 @@ const messageBuckets = new TTLMap();
 const notifiedGuilds = new TTLMap();
 const TIME_WINDOW = 10000;
 const SAME_MSG_THRESHOLD = 3;
-const NOTIFY_COOLDOWN = 30 * 60 * 1000;
+const NOTIFY_COOLDOWN = 60 * 60 * 1000;
+const PASSIVE_MODE = process.env.PASSIVE_MODE == 'true';
 
 export async function handleMessage(message) {
   if (message.author.bot || !message.guild) return;
@@ -22,8 +23,14 @@ export async function handleMessage(message) {
     for (const url of urls) {
       const res = await analyzeUrl(url);
       if (res.malicious > 0 || res.suspicious > 3) {
-        await safeDelete(message);
-        await alertChannel.send(`🚨 Malicious URL from <@${userId}>: ${url} at <#${message.channel.id}>`);
+        if (!PASSIVE_MODE) {
+          await safeDelete(message);
+        }
+        const msg = `${
+          PASSIVE_MODE ? '🔍 [Passive Mode]' : '🚨'
+        } Malicious URL from <@${userId}>: ${url} at <#${message.channel.id}>`;
+
+        await alertChannel.send(msg);
         return;
       }
     }
@@ -39,27 +46,34 @@ export async function handleMessage(message) {
   const { isSpam, messagesToDelete } = detectSpam(recent, userId, content, SAME_MSG_THRESHOLD);
   if (isSpam) {
     const deletedIds = new Set();
-    for (const entry of messagesToDelete) {
-      if (deletedIds.has(entry.messageId)) continue;
-      try {
-        const channel = await message.client.channels.fetch(entry.channelId);
-        if (!channel.isTextBased()) continue;
 
-        const msg = await channel.messages.fetch(entry.messageId);
-        if (msg && msg.deletable) {
-          await msg.delete();
-          deletedIds.add(entry.messageId);
-        }
-      } catch (err) {
-        if (err.code !== 10008) {
-          console.warn(`❌ Error deleting message ${entry.messageId}: ${err.message}`);
+    if (!PASSIVE_MODE) {
+      for (const entry of messagesToDelete) {
+        if (deletedIds.has(entry.messageId)) continue;
+        try {
+          const channel = await message.client.channels.fetch(entry.channelId);
+          if (!channel.isTextBased()) continue;
+
+          const msg = await channel.messages.fetch(entry.messageId);
+          if (msg && msg.deletable) {
+            await msg.delete();
+            deletedIds.add(entry.messageId);
+          }
+        } catch (err) {
+          if (err.code !== 10008) {
+            console.warn(`❌ Error deleting message ${entry.messageId}: ${err.message}`);
+          }
         }
       }
     }
 
     const lastNotified = notifiedGuilds.get(guildId) || 0;
     if (now - lastNotified > NOTIFY_COOLDOWN) {
-      await alertChannel.send(`🚨 Spam detected by <@${userId}> in guild **${message.guild.name}** at <#${message.channel.id}>`);
+      const alertMsg = `${
+        PASSIVE_MODE ? '🔍 [Passive Mode]' : '🚨'
+      } Spam detected by <@${userId}> in guild **${message.guild.name}** at <#${message.channel.id}>`;
+
+      await alertChannel.send(alertMsg);
       notifiedGuilds.set(guildId, now);
     }
   }
